more bug fixes in user controller

+++ src/utils/cloudinary.js
@@ -18,10 +18,16 @@ const uploadOnCloudinary = async (localFilePath) => {
 
     // file has been uploaded succesfully
     // console.log("File is uploaded on cloudinary", response.url);
-    // fs.unlinkSync(localFilePath);
+    // fs.unlink(localFilePath);
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath); // remove the locally saved temporary file as the upload operation got failed
+    // fs.unlinkSync(localFilePath); // remove the locally saved temporary file as the upload operation got failed
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
     return null;
   }
 };
